madfa: Fixed a bug and optimized equivalent-state search.
Made use of the insight in "Constructing Minimal Acyclic Deterministic
Finite Automata" that at each point where states are compared for
equivalence, the order of operations in the algorithm as a whole ensures
that all child states are already unique representatives of their
equivalence classes. As a result, you can simply compare children for
equality rather than needing to recursively check for equivalence.